Updated namespaces, changed structure, refactored, etc.
