+++ src/EasyBlog/Areas/Admin/Models/PostManagementListViewModel.cs
@@ -1,4 +1,4 @@
-namespace EasyBlog.Models.Posts;
+namespace EasyBlog.Areas.Admin.Models;
 
 public class PostManagementListViewModel : BaseViewModel
 {
