+++ src/EasyBlog/Areas/Admin/Validation/PostManagementViewModelValidator.cs
@@ -1,4 +1,5 @@
 // ReSharper disable ClassNeverInstantiated.Global
+
 using FluentValidation;
 
 namespace EasyBlog.Areas.Admin.Validation;
