+++ src/EasyBlog/Areas/Admin/Views/PostsManagement/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model EasyBlog.Models.Posts.PostManagementViewModel
+﻿@model PostManagementViewModel
 
 @{
     ViewData["Title"] = "Post Details";
