+++ src/EasyBlog/Areas/Admin/Views/PostsManagement/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model EasyBlog.Models.Posts.PostManagementListViewModel
+﻿@model PostManagementListViewModel
 
 @{
     ViewData["Title"] = "Posts Management";
