+++ src/EasyBlog/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,1 +1,2 @@
-﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+﻿@using EasyBlog.Areas.Admin.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
