+++ src/EasyBlog/GlobalUsings.cs
@@ -3,12 +3,14 @@
 global using System.Diagnostics;
 global using System.Globalization;
 global using EasyBlog;
+global using EasyBlog.Areas.Admin.Extensions;
+global using EasyBlog.Areas.Admin.Models;
+global using EasyBlog.Areas.Admin.Services;
 global using EasyBlog.Common;
 global using EasyBlog.DataAccess;
 global using EasyBlog.DataAccess.Entities;
 global using EasyBlog.Extensions;
 global using EasyBlog.Models;
-global using EasyBlog.Models.Posts;
 global using EasyBlog.Services;
 global using Microsoft.AspNetCore.Authorization;
 global using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
