+++ src/EasyBlog/Views/Posts/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model EasyBlog.Models.Posts.PostListViewModel
+﻿@model PostListViewModel
 
 @{
     ViewData["Title"] = "Home Page";
