+++ src/EasyBlog/Views/_ViewImports.cshtml
@@ -1,1 +1,2 @@
-﻿@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+﻿@using EasyBlog.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
